src/customer/requirements.txt
Flask==2.3.3
flask-sqlalchemy==2.5.1
PyMySQL==1.0.2
cryptography==41.0.7
flask-migrate==4.0.5
sqlalchemy-utils==0.41.1
sqlalchemy==1.3.24
Flask-JWT-Extended==4.5.3
web3==6.5.0
py-solc-x==2.0.2